Globally searchable false
Whenever i loaded this package i was having problems with the search. I have a big project and i narrowed it down to being this package. Though i had fixed it with assigning the model property but i also needed to make global search false.

Is the search working for anyone?